chore: remove unused files and configurations
- Removed `.github/FUNDING.yml` file as it is no longer needed.
- Removed `.github/ISSUE_TEMPLATE/bug-report---.md` file as it is no longer needed.
- Removed `.github/ISSUE_TEMPLATE/config.yml` file as it is no longer needed.
- Removed `.github/ISSUE_TEMPLATE/feature-request---.md` file as it is no longer needed.
- Removed `.github/workflows/stale.yml` file as it is no longer needed.
- Removed `assets/bmac.png` file as it is no longer needed.
- Removed `assets/logo.png` file as it is no longer needed.

feat: add new files and configurations

- Added `compose.yml` file for docker-compose configuration.
- Added `deployments/gcp/cloudbuild.yaml` file for Google Cloud Build configuration.
- Added `docs/assets/algo724_logo.png` file for documentation.
- Added `docs/index.md` file for documentation.
- Added `mkdocs.yml` file for documentation configuration.
- Added `pyproject.toml` file for Python project configuration.
- Added `tradingview_webhook_bot/__init__.py` file for initializing the TradingView webhook bot.
- Added `tradingview_webhook_bot/config.py` file for configuring the TradingView webhook bot.
- Added `tradingview_webhook_bot/handler.py` file for handling alerts in the TradingView webhook bot.
