--- .github/ISSUE_TEMPLATE/bug-report---.md
@@ -1,22 +1,0 @@
----
-name: "Bug report \U0001F41E"
-about: Create a bug report
-labels: bug
-
----
-
-## Describe the bug
-A clear and concise description of what the bug is.
-
-### Steps to reproduce
-Steps to reproduce the behavior.
-
-### Expected behavior
-A clear and concise description of what you expected to happen.
-
-### Environment
- - OS: [e.g. Arch Linux]
- - Other details that you think may affect.
-
-### Additional context
-Add any other context about the problem here.
