--- .github/ISSUE_TEMPLATE/config.yml
@@ -1,4 +1,0 @@
-contact_links:
-  - name: Question 🙋
-    url: https://github.com/fabston/TradingView-Webhook-Bot/discussions/categories/q-a
-    about: Ask your question in our Discussions Channel
