--- .github/ISSUE_TEMPLATE/feature-request---.md
@@ -1,15 +1,0 @@
----
-name: "Feature request \U0001F680"
-about: Suggest an idea
-labels: enhancement
-
----
-
-## Summary
-Brief explanation of the feature.
-
-### Basic example
-Include a basic example or links here.
-
-### Motivation
-Why are we doing this? What use cases does it support? What is the expected outcome?
