--- .github/workflows/stale.yml
@@ -1,22 +1,0 @@
-name: Mark and close stale issues and pull requests
-
-on:
-  schedule:
-  - cron: "0 0 * * *"
-
-jobs:
-  stale:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/stale@main
-      with:
-        repo-token: ${{ secrets.GITHUB_TOKEN }}
-        stale-issue-message: 'It looks like this has been idle a while, so I am marking it as stale. Remove the label or comment if this issue should remain open.'
-        stale-pr-message: 'It looks like this PR has been idle a while, so I am marking it as stale. Remove the label or comment if this is still being worked on.'
-        stale-issue-label: 'no-activity'
-        stale-pr-label: 'no-activity'
-        exempt-pr-labels: 'help wanted,enhancement'
-        days-before-stale: 30
-        days-before-close: 5
